Rephrase availability of norms and sanctions
According to the model, a norm is defined as a tuple (status,
conditions, issuer, content). However, 'status' should actually be
named 'state' due to its meaning in a norm definition. As this
conflicts with the semantics of normative state in Situated
Artificial Institutions, I've decided to adopt the following:
- enabled - available for activation
- disabled - not available for activation
- active - conditions hold in the current state
- inactive - conditions don't hold in the current state

As a result, the availability of a norm/sanction is now passed as
an XML attribute in the De Jure specification and the 'status'
element has been removed.

Signed-off-by: Igor C. A. de Lima <igorcadelima@gmail.com>