Rename SchemeBoard#rambos_updateRolePlayers
Previously, the SchemeBoard#updateRolePlayers method  was private
in Moise. Because of that, I had to create another method to
make things work properly. I called it rambos_updateRolePlayers.

A recent commit in Moise's repository, changed the method's
visibility to protected. Because of that, I am renaming the
method to updateRolePlayers.

See https://github.com/moise-lang/moise/commit/88c1574

Signed-off-by: Igor C. A. de Lima <igorcadelima@gmail.com>